# src/main/java/com/cft/rest/api/book/UpdateBookRequest.java
package com.cft.rest.api.book;

import java.util.List;

import com.fasterxml.jackson.annotation.JsonProperty;

public class UpdateBookRequest {

	private final String title;
	private final Integer yearIssued;
	private final Integer numberOfCopies;
	private final List<String> authors;
	
	public UpdateBookRequest(
			@JsonProperty(value = "title") String title, 
			@JsonProperty(value = "yearIssued") Integer yearIssued, 
			@JsonProperty(value = "numberOfCopies") Integer numberOfCopies, 
			@JsonProperty(value = "authors") List<String> authors) {
		
		this.title = title;
		this.yearIssued = yearIssued;
		this.numberOfCopies = numberOfCopies;
		this.authors = authors;
	}

	public String getTitle() {
		return title;
	}

	public Integer getYearIssued() {
		return yearIssued;
	}

	public Integer getNumberOfCopies() {
		return numberOfCopies;
	}

	public List<String> getAuthors() {
		return authors;
	}
	
}